pyproject.toml
[project]
name = "langgraph-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorlog>=6.9.0",
    "fastapi>=0.115.12",
    "gradio>=5.29.1",
    "langchain>=0.3.25",
    "langgraph>=0.4.5",
    "langsmith>=0.3.42",
    "notebook>=7.4.2",
    "openai>=1.79.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "serpapi>=0.1.5",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
